part of medicine api

+++ MedicineManageProject/Model/DISCOUNT.cs
@@ -8,8 +8,9 @@ namespace MedicineManageProject.Model
 {
     public class DISCOUNT
     {
-        [SugarColumn(IsPrimaryKey = true,IsIdentity =true)]
-        public Decimal DISCOUNT_ID { get; set; }
+        [SugarColumn(IsPrimaryKey = true,IsIdentity = true)]
+        public int DISCOUNT_ID { get; set; }
+
         public Decimal DISCOUNT_PRICE { get; set; }
         public DateTime START_TIME { get; set; }
         public DateTime END_TIME { get; set; }
